Add Booking and BookingResource

+++ DotNetAngularApp/Models/TimeSlot.cs
@@ -7,12 +7,10 @@ namespace DotNetAngularApp.Models
     [Table("TimeSlots")]
     public class TimeSlot
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
-        //[DisplayFormat(DataFormatString = "{0:hh}:{0:mm}", ApplyFormatInEditMode = true)]
         public DateTime StartTime { get; set; }
 
-        //[DisplayFormat(DataFormatString = "{0:hh}:{0:mm}", ApplyFormatInEditMode = true)]
         public DateTime EndTime { get; set; }
 
         //public bool IsBooked { get; set; }
